+++ packages/provider-queries/src/recommended-articles.js
@@ -1,0 +1,4 @@
+import { addTypenameToDocument } from "apollo-utilities";
+import recommendations from "./recommended_articles.graphql";
+
+export default addTypenameToDocument(recommendations);