+++ packages/provider/src/recommended-articles.js
@@ -1,0 +1,4 @@
+import { recommendations } from "@times-components/provider-queries";
+import connectGraphql from "./connect";
+
+export default connectGraphql(recommendations);