+++ packages/key-facts/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.10.23](https://github.com/newsuk/times-components/compare/@times-components/key-facts@2.10.22...@times-components/key-facts@2.10.23) (2022-05-25)
+
+**Note:** Version bump only for package @times-components/key-facts
+
+
+
+
+
 ## [2.10.22](https://github.com/newsuk/times-components/compare/@times-components/key-facts@2.10.21...@times-components/key-facts@2.10.22) (2022-05-16)
 
 **Note:** Version bump only for package @times-components/key-facts