+++ packages/related-articles/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.8.15](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.8.14...@times-components/related-articles@6.8.15) (2022-05-25)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
 ## [6.8.14](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.8.13...@times-components/related-articles@6.8.14) (2022-05-16)
 
 **Note:** Version bump only for package @times-components/related-articles