+++ packages/save-star-web/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.110](https://github.com/newsuk/times-components/compare/@times-components/save-star-web@0.6.109...@times-components/save-star-web@0.6.110) (2022-05-25)
+
+**Note:** Version bump only for package @times-components/save-star-web
+
+
+
+
+
 ## [0.6.109](https://github.com/newsuk/times-components/compare/@times-components/save-star-web@0.6.108...@times-components/save-star-web@0.6.109) (2022-05-16)
 
 **Note:** Version bump only for package @times-components/save-star-web